#106 FEAT : Query와 관련된 메서드를 RoomQueryService로 이전
